src/AI func/PhotoComparison/requirements.txt
openai
flask
python-dotenv
requests
